backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
starlette>=0.35.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
openai>=1.12.0
httpx>=0.26.0
sentence-transformers>=2.3.1
torch>=2.2.0
transformers>=4.37.0
numpy>=1.26.0
PyPDF2
faiss-cpu
rank_bm25
google-generativeai
python-dotenv>=1.0.1
aiofiles>=23.2.1
tenacity>=8.2.3
psutil>=5.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
scikit-learn>=1.3.0